Separate Ecosia code into separate files
