+++ firefox-ios/Storage/RustKeychain.swift
@@ -345,27 +345,6 @@ public final class RustKeychain: KeychainProtocol {
         return keychainQueryDictionary
     }
 
-    // Ecosia: queries keychain items written by the legacy MZKeychainWrapper (pre-v147).
-    // Uses kSecAttrAccount as a plain String (matching MZKeychainWrapper's storage format) and
-    // omits kSecAttrGeneric so it finds the original item rather than any RustKeychain item.
-    public func legacyDataForKey(_ key: String) -> Data? {
-        var query: [String: Any] = [
-            kSecClass as String: kSecClassGenericPassword,
-            kSecAttrService as String: self.serviceName,
-            kSecAttrAccount as String: key,         // String, not Data — matches MZKeychainWrapper
-            kSecAttrSynchronizable as String: false,
-            kSecReturnData as String: true,
-            kSecMatchLimit as String: kSecMatchLimitOne
-        ]
-        if let accessGroup = self.accessGroup {
-            query[kSecAttrAccessGroup as String] = accessGroup
-        }
-        var result: AnyObject?
-        let status = SecItemCopyMatching(query as CFDictionary, &result)
-        guard status == errSecSuccess, let data = result as? Data else { return nil }
-        return data
-    }
-
     private func logErrorFromStatus(_ status: OSStatus, errMsg: String) {
         guard status != errSecSuccess else {
             return
